Renames run-create-app -> create-mock-app
