+++ test/stateful-stateless.js
@@ -3,14 +3,14 @@ const path = require('path');
 const test = require('ava');
 
 const getBinPath = require('./utils/get-bin-path');
-const runCreateApp = require('./utils/run-create-app');
+const createMockApp = require('./utils/create-mock-app');
 const runWithPrompt = require('./utils/run-with-prompt');
 
 const template = async (t, command, answers, args = '') => {
   t.plan(2);
 
   const componentName = answers[0] || args;
-  const buildPath = await runCreateApp();
+  const buildPath = await createMockApp();
 
   // Convert to stateful
   await runWithPrompt(
